requirements.txt
flask==3.0.0
pandas==2.2.0
numpy==1.26.0
python-dotenv==1.0.0
marshmallow==3.20.1
pytest==7.4.4
flask-cors==4.0.0
datetime
requests